using double

+++ alura/guessGame.c
@@ -10,9 +10,11 @@ int main() {
     int numerosecreto = 42;
     
     int chute;
+
+
     int tentativas = 1;
 
-    int pontos = 1000;
+    double pontos = 1000;
 
     while(1) {
         printf("Tentativa %d\n", tentativas);
@@ -41,13 +43,13 @@ int main() {
         }
 
         tentativas++;
-        int pontosperdidos = (chute - numerosecreto) / 2;
+        double pontosperdidos = (chute - numerosecreto) / 2.0;
         pontos = pontos - pontosperdidos;
     }
 
     printf("Fim de jogo!\n");
     printf("Você acertou em %d tentativas!", tentativas);
-    printf("Pontuação: %d \n", pontos);
+    printf("Pontuação: %f \n", pontos);
